Created first version of the code base
First limited but working version. The meta data is not properly
serialized to and from the file system. The file system can currently
only span a single sector as the tree routines cannot handle splitting
of nodes or data in dedicated sectors. The SubTreeTransaction
implementation is not complete but works for small tests. Tests and
documentation is missing. A single file system is created when the system
is started with hard coded UUIDs. The system is not persistent.